requirements.txt
apache-airflow==2.9.3
google-cloud-bigquery==3.41.0
requests==2.32.3
pytest==8.3.5
